new convention of parsing responses
